pyproject.toml
[project]
name = "pdfsvg-calibrator"
version = "0.1.0"
description = "Axis-aligned PDF→SVG scale calibration with numbered line verification."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "lxml>=5.0",
  "PyMuPDF>=1.24",   # fitz
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7"
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "black", "mypy", "pre-commit"]

[project.scripts]
pdfsvg-calibrate = "pdfsvg_calibrator.cli:app"
